Add files via upload
A target audience profile is simply a particular community of clients most likely to respond to your
advertisements, goods and services positively. At this point we know the overarching meaning of the target
audience: a community of people you want to sell something to. These are the website users that we consider
we want to attract and captivate.
Target audience research will also be focused on particular variables such as place, age, income, and so on.
Basically, an overview of the target audience provides us with guidance for our messaging so that we can
create deeper consumer relationships.